+++ core/parse.c
@@ -10,7 +10,7 @@
 #include "divesite.h"
 #include "subsurface-string.h"
 #include "parse.h"
-#include "divelist.h"
+#include "trip.h"
 #include "device.h"
 #include "gettext.h"
