+++ core/save-git.c
@@ -19,7 +19,7 @@
 
 #include "divesite.h"
 #include "subsurface-string.h"
-#include "divelist.h"
+#include "trip.h"
 #include "device.h"
 #include "membuffer.h"
 #include "git-access.h"
