+++ core/save-xml.c
@@ -15,7 +15,7 @@
 
 #include "divesite.h"
 #include "subsurface-string.h"
-#include "divelist.h"
+#include "trip.h"
 #include "device.h"
 #include "membuffer.h"
 #include "strndup.h"
