+++ core/statistics.c
@@ -12,7 +12,7 @@
 
 #include "dive.h"
 #include "display.h"
-#include "divelist.h"
+#include "trip.h"
 #include "statistics.h"
 #include "units.h"
