+++ desktop-widgets/divelogimportdialog.cpp
@@ -12,6 +12,7 @@
 #include <QPainter>
 #include "core/qthelper.h"
 #include "core/divesite.h"
+#include "core/trip.h"
 #include "core/import-csv.h"
 
 static QString subsurface_mimedata = "subsurface/csvcolumns";
