+++ export-html.cpp
@@ -10,6 +10,7 @@
 #include "core/qthelper.h"
 #include "core/file.h"
 #include "core/divesite.h"
+#include "core/trip.h"
 #include "core/save-html.h"
 #include <stdio.h>
 #include "git2.h"
