+++ qt-models/filtermodels.cpp
@@ -4,6 +4,7 @@
 #include "core/display.h"
 #include "core/qthelper.h"
 #include "core/divesite.h"
+#include "core/trip.h"
 #include "core/subsurface-string.h"
 #include "core/subsurface-qt/DiveListNotifier.h"
 #include "qt-models/divetripmodel.h"
